+++ gendata/gendata.sh
@@ -1,0 +1,3 @@
+#!/bin/zsh
+javac -d ./jv/ ./src/*.java
+java -cp jv GenData